autherization and authentication

+++ mernApp/backend/controller/goalController.js
@@ -1,12 +1,14 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../model/goalmodel');
+const User = require('../model/userModel');
+
 
 // @Des GetGoals
 // @route GET /api/goals
 //@access private
 const getGoals = asyncHandler(async (req, res) => {
 
-     const goals = await Goal.find()
+     const goals = await Goal.find({ user: req.user.id})
     res.status(200).json(goals)
 })
 
@@ -15,14 +17,15 @@ const getGoals = asyncHandler(async (req, res) => {
 // @route POST /api/goals
 //@access private
 const setGoals = asyncHandler(async (req, res) => {
-
+    
     if(!req.body.text) {
         res.status(400)
         throw new Error("please add a text field")
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -31,6 +34,10 @@ const setGoals = asyncHandler(async (req, res) => {
 // @Des updateGoals
 // @route PUT /api/goals/:id
 //@access private
+
+
+
+
 const updateGoals = asyncHandler(async (req, res) => {
 
     const goal = await Goal.findById(req.params.id);
@@ -39,6 +46,21 @@ const updateGoals = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal not found')
      }
+
+ const user = await User.findById(req.user.id)
+
+// Check for user
+ if(!user) {
+    res.status(401)
+    throw new Error('User not found')
+ }
+
+ // Make sure the logged user match the goal user
+ if(goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
+ }
+
 
      const updateGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
@@ -59,6 +81,22 @@ const deleteGoals = asyncHandler(async (req, res) => {
        res.status(400)
        throw new Error('Goal not found')
     }
+
+
+    const user = await User.findById(req.user.id)
+
+// Check for user
+ if(!user) {
+    res.status(401)
+    throw new Error('User not found')
+ }
+
+ // Make sure the logged user match the goal user
+ if(goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
+ }
+ 
     await Goal.deleteOne({ _id: req.params.id });
         res.status(200).json({id: req.params.id})
